update dao changed to sigleton and others..
