+++ Library_System/src/model/dao/OverdueDao.java
@@ -1,18 +1,30 @@
-package model;
-// 
+package model.dao;
+
 import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.ArrayList;
-//
+
 import db.DBConnector;
+import model.dto.Overdue;
+
 
 public class OverdueDao {
-	    
-	public ArrayList<OverdueDto> list() throws SQLException {
+	private static OverdueDao instance;
+	ArrayList<Overdue> overdueList;
+	
+	private OverdueDao() {}
+	
+	public static OverdueDao getInstance() {
+		if(instance == null) {
+			instance = new OverdueDao();
+		}
+		return instance;
+	}
+	public ArrayList<Overdue> listAllOverdue() {
 
-		ArrayList<OverdueDto> overdueList = new ArrayList<OverdueDto>();
+		overdueList = new ArrayList<>();
 		String sql = "SELECT * FROM overdue";
 
 		try (
@@ -27,7 +39,7 @@ public class OverdueDao {
 				int book_id = rs.getInt("book_id");
 				int overdue_period = rs.getInt("overdue_period");
 
-				OverdueDto loan = new OverdueDto(loan_num, book_id, student_num, overdue_period);
+				Overdue loan = new Overdue(loan_num, book_id, student_num, overdue_period);
 
 				overdueList.add(loan);
