+++ Library_System/src/view/component/DefaultBtn.java
@@ -1,0 +1,12 @@
+package view.component;
+
+import java.awt.Dimension;
+
+import javax.swing.JButton;
+
+public class DefaultBtn extends JButton{
+	public DefaultBtn() {
+		setPreferredSize(new Dimension(100,40));
+	}
+
+}
